API Documentation

Docstrings for AutoSysimages.jl interface members can be accessed through Julia's built-in documentation system or in the list below.

Contents

API Documentation

Contents

Index

Functions

Index

AutoSysimages.active_dir

AutoSysimages.add

AutoSysimages.build_sysimage

AutoSysimages.julia_args

AutoSysimages.latest_sysimage

AutoSysimages.packages_to_include

AutoSysimages.preferences_path

AutoSysimages.remove

AutoSysimages.remove_old_sysimages

AutoSysimages.set_packages

AutoSysimages.start

AutoSysimages.status

Functions

AutoSysimages.start — Function

start()

Starts AutoSysimages package. It's usually called by start.jl file; but it can be called manually as well.

AutoSysimages.latest_sysimage — Function

latest_sysimage()

Return the path to the latest system image produced by AutoSysimages, or nothing if no such image exits.

AutoSysimages.julia_args — Function

julia_args()

Get Julia arguments for running AutoSysimages:

"-J [sysimage]" - sets the latest_sysimage(), if it exits,

"-L [@__DIR__]/start.jl" - starts AutoSysimages automatically.

AutoSysimages.build_sysimage — Function

build_sysimage(background::Bool = false)

Build new system image (in background) for the current project including snooped precompiles.

AutoSysimages.remove_old_sysimages — Function

remove_old_sysimages()

Remove old sysimages for the current project (active_dir()).

AutoSysimages.packages_to_include — Function

packages_to_include()::Set{String}

Get list of packages to be included into sysimage. It is determined based on "include" or "exclude" options save by Preferences.jl in LocalPreferences.toml file next to the currently-active project. Notice dev packages are excluded unless they are in include list.

AutoSysimages.set_packages — Function

set_packages()

Ask the user to choose which packages to include into the sysimage.

AutoSysimages.status — Function

status()

Print list of packages to be included into sysimage determined by packages_to_include.

AutoSysimages.add — Function

add(package::String)

Set package to be included into the system image.

AutoSysimages.remove — Function

remove(package::String)

Set package to be excluded into the system image.

AutoSysimages.active_dir — Function

active_dir()

Get directory where the sysimage and precompiles are stored for the current project. The directory is created if it doesn't exist yet.

AutoSysimages.preferences_path — Function

preferences_path()

Get the file with preferences for the active project (active_dir()). Preferences are stored in SysimagePreferences.toml next to the current Project.toml file.